requirements.txt
aioredis==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.1.0
constantly==15.1.0
cryptography==41.0.1
daphne==4.0.0
defusedxml==0.7.1
Django==4.2.2
django-allauth==0.54.0
django-filter==23.2
djangorestframework==3.14.0
hiredis==2.2.3
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
msgpack==1.0.5
oauthlib==3.2.2
Pillow==10.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
pyOpenSSL==23.2.0
python3-openid==3.2.0
pytz==2023.3
redis==4.6.0
requests==2.31.0
requests-oauthlib==1.3.1
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==22.10.0
twisted-iocpsupport==1.0.3
txaio==23.1.1
typing_extensions==4.7.0
tzdata==2023.3
urllib3==2.0.3
zope.interface==6.0
